+++ core/src/main/scala/json/schema/discriminator.scala
@@ -1,0 +1,5 @@
+package json.schema
+
+import scala.annotation.StaticAnnotation
+
+final case class discriminator(field: String = "type", phantom: Boolean = true) extends StaticAnnotation