+++ core/src/main/scala/json/schema/discriminatorKey.scala
@@ -1,0 +1,5 @@
+package json.schema
+
+import scala.annotation.StaticAnnotation
+
+final case class discriminatorKey(label: String) extends StaticAnnotation