+++ macros/src/main/scala/com/github/andyglow/jsonschema/SchemaTypes.scala
@@ -18,16 +18,12 @@ private[jsonschema] trait SchemaTypes { this: UContext with UCommons =>
 
     final def tree: Tree = {
 
-      val withValidationsApplied = if (extra.validations.nonEmpty) q"$prefix.withValidation(..${extra.validations})" else prefix
+      val withValidationsApplied  = if (extra.validations.nonEmpty) q"$prefix.withValidation(..${extra.validations})" else prefix
+      val withTitle               = extra.title.fold(withValidationsApplied) { x => q"$withValidationsApplied.withTitle($x)" }
+      val withDescription         = extra.description.fold(withTitle) { x => q"$withTitle.withDescription($x)" }
+      val withDiscriminationKey   = extra.discriminationKey.fold(withDescription) { x => q"$withDescription.withDiscriminationKey($x)" }
 
-      val decorated = (extra.title, extra.description) match {
-        case (Some(title), Some(description)) => q"$withValidationsApplied.duplicate(title = Some($title), description = Some($description))"
-        case (Some(title), None)              => q"$withValidationsApplied.duplicate(title = Some($title))"
-        case (None, Some(description))        => q"$withValidationsApplied.duplicate(description = Some($description))"
-        case (None, None)                     => withValidationsApplied
-      }
-
-      decorated
+      withDiscriminationKey
     }
   }
 
@@ -35,7 +31,8 @@ private[jsonschema] trait SchemaTypes { this: UContext with UCommons =>
     final case class Extra(
       validations: Seq[Tree] = Seq.empty,
       title: Option[String] = None,
-      description: Option[String] = None)
+      description: Option[String] = None,
+      discriminationKey: Option[String] = None)
 
     final case class Bool(extra: Extra = Extra()) extends SchemaType { type Self = Bool; def prefix = q"${N.Schema}.`boolean`()"; val tpe = typeOf[Boolean]; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class Integer(extra: Extra = Extra()) extends SchemaType { type Self = Integer; def prefix = q"${N.Schema}.`integer`()"; val tpe = typeOf[Int]; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
@@ -45,7 +42,7 @@ private[jsonschema] trait SchemaTypes { this: UContext with UCommons =>
     final case class Dict(keyTpe: Type, valueTpe: Type, containerTpe: Type, valueSchema: SchemaType, extra: Extra = Extra()) extends SchemaType { type Self = Dict; def prefix = q"${N.Schema}.`dictionary`[$keyTpe, $valueTpe, $containerTpe](${valueSchema.tree})"; val tpe = appliedType(containerTpe, List(keyTpe, valueTpe)); def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class Obj(tpe: Type, fields: Seq[Obj.Field], extra: Extra = Extra()) extends SchemaType { type Self = Obj; def prefix = q"${N.Schema}.`object`[$tpe](..${fields map { _.tree }})"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class Enum(tpe: Type, values: Seq[Tree], /*duplicates values if created from strings*/names: Option[Seq[String]], extra: Extra = Extra()) extends SchemaType { type Self = Enum; def prefix = q"${N.Schema}.`enum`.of[$tpe](..$values)"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
-    final case class OneOf(tpe: Type, memberSchema: Seq[SchemaType], extra: Extra = Extra()) extends SchemaType { type Self = OneOf; def prefix = q"${N.Schema}.`oneof`[$tpe](Set(..${memberSchema map { _.tree }}))"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
+    final case class OneOf(tpe: Type, memberSchema: Seq[SchemaType], discriminatorField: Option[String], extra: Extra = Extra()) extends SchemaType { type Self = OneOf; def prefix = q"${N.Schema}.`oneof`[$tpe](Set(..${memberSchema map { _.tree }}), $discriminatorField)"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class AllOf(tpe: Type, memberSchema: Seq[SchemaType], extra: Extra = Extra()) extends SchemaType { type Self = AllOf; def prefix = q"${N.Schema}.`allof`[$tpe](Set(..${memberSchema map { _.tree }}))"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class Not(tpe: Type, schema: SchemaType, extra: Extra = Extra()) extends SchemaType { type Self = Not; def prefix = q"${N.Schema}.`not`[$tpe](${schema.tree})"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
     final case class ValueClass(tpe: Type, innerTpe: Type, schema: SchemaType, extra: Extra = Extra()) extends SchemaType { type Self = ValueClass; def prefix = q"${N.Schema}.`value-class`[$tpe, $innerTpe](${schema.tree})"; def withExtra(x: SchemaType.Extra) = copy(extra = x) }
@@ -58,9 +55,10 @@ private[jsonschema] trait SchemaTypes { this: UContext with UCommons =>
         val description: Option[String]
         def tree: Tree = description.fold(prefix)(d => q"$prefix.withDescription(Some($d))")
         def mapSchema(fn: SchemaType => SchemaType): Field
+        def name: String
       }
       final object Field {
-        final case class Apply(tpe: Type, name: Tree, schema: SchemaType, required: Option[Tree], default: Option[Tree], description: Option[String]) extends Field {
+        final case class Apply(tpe: Type, name: String, schema: SchemaType, required: Option[Tree], default: Option[Tree], description: Option[String]) extends Field {
           def prefix = (required, default) match {
             case (Some(required), Some(default))  => q"${N.Schema}.`object`.Field[$tpe]($name, ${schema.tree}, $required, $default)"
             case (None, Some(default))            => q"${N.Schema}.`object`.Field[$tpe]($name, ${schema.tree}, false, $default)"
@@ -69,7 +67,7 @@ private[jsonschema] trait SchemaTypes { this: UContext with UCommons =>
           }
           def mapSchema(fn: SchemaType => SchemaType): Field = copy(schema = fn(schema))
         }
-        final case class FromJson(tpe: Type, name: Tree, schema: SchemaType, required: Tree, default: Tree, description: Option[String]) extends Field {
+        final case class FromJson(tpe: Type, name: String, schema: SchemaType, required: Tree, default: Tree, description: Option[String]) extends Field {
           def prefix = q"${N.Schema}.`object`.Field.fromJson[$tpe]($name, ${schema.tree}, $required, $default)"
           def mapSchema(fn: SchemaType => SchemaType): Field = copy(schema = fn(schema))
         }